+++ app/lib/package/backend.dart
@@ -901,6 +901,20 @@ class PackageBackend {
         }
       }
 
+      // check existences of referenced packages
+      final dependencies = <String>{
+        ...pubspec.dependencies.keys,
+      };
+      for (final name in dependencies) {
+        if (isSoftRemoved(name)) {
+          continue;
+        }
+        if (await isPackageVisible(name)) {
+          continue;
+        }
+        throw PackageRejectedException.dependencyDoesNotExists(name);
+      }
+
       sw.reset();
       final entities = await _createUploadEntities(db, agent, archive,
           sha256Hash: sha256Hash);